requirements.txt
matplotlib
numpy
scipy
scikit-learn
beautifulsoup4
tweet-preprocessor
pandas
nltk
